+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## [`master`](https://github.com/elastic/eui/tree/master)
 
 - Added conditional rendering of the title element in `EuiCallOut` to avoid usage of additional space caused by the rendered `<div>` element ([#3549](https://github.com/elastic/eui/pull/3549))
+- Added new prop to `EuiForm` to allow conditional rendering of error callout ([#3585](https://github.com/elastic/eui/pull/3585))
 
 **Bug fixes**